build(tests): reorganize tests structure
